KAN-41 : Rename: User chore
