+++ src/main/java/capstone/relation/api/auth/config/CorsConfig.java
@@ -12,7 +12,6 @@ public class CorsConfig {
 
 	@Bean
 	public CorsFilter corsFilter() {
-		System.out.println("CORS 필터 등록");
 		UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
 		CorsConfiguration config = new CorsConfiguration();
 		config.setAllowCredentials(true); // 내 서버가 응답을 할 때 json을 자바스크립트에서 처리할 수 있게 할지를 설정
