+++ src/main/java/capstone/relation/api/auth/jwt/refreshToken/RefreshToken.java
@@ -2,7 +2,7 @@ package capstone.relation.api.auth.jwt.refreshToken;
 
 import java.util.Date;
 
-import capstone.relation.api.auth.domain.User;
+import capstone.relation.user.domain.User;
 
 public record RefreshToken(
 	String key,
