+++ src/main/java/capstone/relation/api/auth/oauth/provider/OAuthUserProvider.java
@@ -1,6 +1,6 @@
 package capstone.relation.api.auth.oauth.provider;
 
-import capstone.relation.api.auth.domain.User;
+import capstone.relation.user.domain.User;
 
 public interface OAuthUserProvider {
 	User getUser(String authorization);
