+++ src/main/java/capstone/relation/api/auth/oauth/provider/kakao/KakaoUserProvider.java
@@ -9,12 +9,12 @@ import org.springframework.stereotype.Component;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import capstone.relation.api.auth.domain.User;
 import capstone.relation.api.auth.jwt.JwtProperties;
 import capstone.relation.api.auth.oauth.provider.OAuthUserProvider;
 import capstone.relation.api.auth.oauth.provider.kakao.client.KakaoOAuthClient;
 import capstone.relation.api.auth.oauth.provider.kakao.client.KakaoTokenClient;
 import capstone.relation.api.auth.oauth.provider.kakao.response.KakaoOAuth2Response;
+import capstone.relation.user.domain.User;
 import lombok.RequiredArgsConstructor;
 
 @Component
@@ -23,7 +23,7 @@ public class KakaoUserProvider implements OAuthUserProvider {
 	private final KakaoOAuthClient kakaoClient;
 	private final KakaoTokenClient kakaoTokenClient;
 	private final JwtProperties jwtProperties;
-	
+
 	@Value("${oauth2.kakao.client-id}")
 	private String clientId;
