+++ src/main/java/capstone/relation/api/auth/oauth/provider/naver/NaverUserProvider.java
@@ -10,12 +10,12 @@ import org.springframework.stereotype.Component;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import capstone.relation.api.auth.domain.User;
 import capstone.relation.api.auth.jwt.JwtProperties;
 import capstone.relation.api.auth.oauth.provider.OAuthUserProvider;
 import capstone.relation.api.auth.oauth.provider.naver.client.NaverOAuthClient;
 import capstone.relation.api.auth.oauth.provider.naver.client.NaverTokenClient;
 import capstone.relation.api.auth.oauth.provider.naver.response.NaverOAuth2Response;
+import capstone.relation.user.domain.User;
 import lombok.RequiredArgsConstructor;
 
 @Component
@@ -24,7 +24,7 @@ public class NaverUserProvider implements OAuthUserProvider {
 	private final NaverOAuthClient naverOAuthClient;
 	private final NaverTokenClient naverTokenClient;
 	private final JwtProperties jwtProperties;
-	
+
 	@Value("${oauth2.naver.client-id}")
 	private String clientId;
 	@Value("${oauth2.naver.client-secret}")
