+++ src/main/java/capstone/relation/api/auth/oauth/provider/naver/response/NaverOAuth2Response.java
@@ -1,14 +1,14 @@
 package capstone.relation.api.auth.oauth.provider.naver.response;
 
-import capstone.relation.api.auth.domain.Role;
-import capstone.relation.api.auth.domain.User;
+import capstone.relation.user.domain.Role;
+import capstone.relation.user.domain.User;
 
 public record NaverOAuth2Response(String resultCode, String message, UserInfo response) {
 
 	public User toEntity() {
 		return User.builder()
 			.email(response.email())
-			.username(response.name())
+			.userName(response.name())
 			.profileImage(response.profile_image())
 			.provider("naver")
 			.role(Role.USER)
